requirements.txt
django
django-compressor
django-extensions
django-grappelli
django-libsass
django-model-utils
django-redis
psycopg2
redis
elasticsearch>=5.0.0,<6.0.0


django-kdl-ldap
django-haystack